third_party/blink/web_tests/external/wpt/css/css-masking/parsing/mask-valid.sub-expected.txt

This is a testharness.js-based test.
Found 55 tests; 30 PASS, 25 FAIL, 0 TIMEOUT, 0 NOTRUN.
[PASS] e.style['mask'] = "none" should set the property value
[PASS] e.style['mask'] = "linear-gradient(to left bottom, red, blue)" should set the property value
[PASS] e.style['mask'] = "linear-gradient(to left bottom, red, blue) luminance" should set the property value
[PASS] e.style['mask'] = "url(\"https://web-platform.test/\")" should set the property value
[PASS] e.style['mask'] = "linear-gradient(to left bottom, red, blue) 1px 2px" should set the property value
[PASS] e.style['mask'] = "url(\"https://web-platform.test/\") 1px 2px / contain" should set the property value
[PASS] e.style['mask'] = "repeat-y" should set the property value
[PASS] e.style['mask'] = "border-box" should set the property value
[PASS] e.style['mask'] = "linear-gradient(to left bottom, red, blue) padding-box" should set the property value
[PASS] e.style['mask'] = "content-box" should set the property value
[PASS] e.style['mask'] = "url(\"https://web-platform.test/\") fill-box" should set the property value
[PASS] e.style['mask'] = "linear-gradient(to left bottom, red, blue) stroke-box" should set the property value
[PASS] e.style['mask'] = "view-box" should set the property value
[PASS] e.style['mask'] = "no-clip" should set the property value
[PASS] e.style['mask'] = "url(\"https://web-platform.test/\") add" should set the property value
[PASS] e.style['mask'] = "subtract" should set the property value
[PASS] e.style['mask'] = "url(\"https://web-platform.test/\") intersect" should set the property value
[PASS] e.style['mask'] = "linear-gradient(to left bottom, red, blue) exclude" should set the property value
[PASS] e.style['mask'] = "alpha" should set the property value
[PASS] e.style['mask'] = "url(\"https://web-platform.test/\") alpha" should set the property value
[PASS] e.style['mask'] = "border-box border-box" should set the property value
[PASS] e.style['mask'] = "content-box content-box" should set the property value
[PASS] e.style['mask'] = "border-box content-box" should set the property value
[PASS] e.style['mask'] = "border-box no-clip" should set the property value
[PASS] e.style['mask'] = "intersect no-clip space round 1px 2px / contain stroke-box linear-gradient(to left bottom, red, blue) luminance" should set the property value
[PASS] e.style['mask'] = "intersect no-clip space round 1px 2px / contain view-box, stroke-box linear-gradient(to left bottom, red, blue) luminance" should set the property value
[FAIL] e.style['mask'] = "none alpha" should set the property value
  assert_equals: serialization should be canonical expected "alpha" but got "none alpha"
[FAIL] e.style['mask'] = "none" should set mask-border-outset
  assert_equals: mask-border-outset should be canonical expected (string) "0" but got (undefined) undefined
[FAIL] e.style['mask'] = "none" should set mask-border-repeat
  assert_equals: mask-border-repeat should be canonical expected (string) "stretch" but got (undefined) undefined
[FAIL] e.style['mask'] = "none" should set mask-border-slice
  assert_equals: mask-border-slice should be canonical expected (string) "0" but got (undefined) undefined
[FAIL] e.style['mask'] = "none" should set mask-border-source
  assert_equals: mask-border-source should be canonical expected (string) "none" but got (undefined) undefined
[FAIL] e.style['mask'] = "none" should set mask-border-width
  assert_equals: mask-border-width should be canonical expected (string) "auto" but got (undefined) undefined
[FAIL] e.style['mask'] = "none" should set mask-clip
  assert_equals: mask-clip should be canonical expected "border-box" but got "initial"
[FAIL] e.style['mask'] = "none" should set mask-composite
  assert_equals: mask-composite should be canonical expected "add" but got "initial"
[PASS] e.style['mask'] = "none" should set mask-image
[FAIL] e.style['mask'] = "none" should set mask-mode
  assert_equals: mask-mode should be canonical expected "match-source" but got "initial"
[FAIL] e.style['mask'] = "none" should set mask-origin
  assert_equals: mask-origin should be canonical expected "border-box" but got "initial"
[FAIL] e.style['mask'] = "none" should set mask-position
  assert_equals: mask-position should be canonical expected "0% 0%" but got "initial"
[FAIL] e.style['mask'] = "none" should set mask-repeat
  assert_equals: mask-repeat should be canonical expected "repeat" but got "initial"
[FAIL] e.style['mask'] = "none" should set mask-size
  assert_equals: mask-size should be canonical expected "auto" but got "initial"
[PASS] e.style['mask'] = "none" should not set unrelated longhands
[FAIL] e.style['mask'] = "none, linear-gradient(to left bottom, red, blue) padding-box" should set mask-border-outset
  assert_equals: mask-border-outset should be canonical expected (string) "0" but got (undefined) undefined
[FAIL] e.style['mask'] = "none, linear-gradient(to left bottom, red, blue) padding-box" should set mask-border-repeat
  assert_equals: mask-border-repeat should be canonical expected (string) "stretch" but got (undefined) undefined
[FAIL] e.style['mask'] = "none, linear-gradient(to left bottom, red, blue) padding-box" should set mask-border-slice
  assert_equals: mask-border-slice should be canonical expected (string) "0" but got (undefined) undefined
[FAIL] e.style['mask'] = "none, linear-gradient(to left bottom, red, blue) padding-box" should set mask-border-source
  assert_equals: mask-border-source should be canonical expected (string) "none" but got (undefined) undefined
[FAIL] e.style['mask'] = "none, linear-gradient(to left bottom, red, blue) padding-box" should set mask-border-width
  assert_equals: mask-border-width should be canonical expected (string) "auto" but got (undefined) undefined
[FAIL] e.style['mask'] = "none, linear-gradient(to left bottom, red, blue) padding-box" should set mask-clip
  assert_equals: mask-clip should be canonical expected "border-box, padding-box" but got "initial, padding-box"
[FAIL] e.style['mask'] = "none, linear-gradient(to left bottom, red, blue) padding-box" should set mask-composite
  assert_equals: mask-composite should be canonical expected "add, add" but got "initial, initial"
[PASS] e.style['mask'] = "none, linear-gradient(to left bottom, red, blue) padding-box" should set mask-image
[FAIL] e.style['mask'] = "none, linear-gradient(to left bottom, red, blue) padding-box" should set mask-mode
  assert_equals: mask-mode should be canonical expected "match-source, match-source" but got "initial, initial"
[FAIL] e.style['mask'] = "none, linear-gradient(to left bottom, red, blue) padding-box" should set mask-origin
  assert_equals: mask-origin should be canonical expected "border-box, padding-box" but got "initial, padding-box"
[FAIL] e.style['mask'] = "none, linear-gradient(to left bottom, red, blue) padding-box" should set mask-position
  assert_equals: mask-position should be canonical expected "0% 0%, 0% 0%" but got ""
[FAIL] e.style['mask'] = "none, linear-gradient(to left bottom, red, blue) padding-box" should set mask-repeat
  assert_equals: mask-repeat should be canonical expected "repeat, repeat" but got "initial, initial"
[FAIL] e.style['mask'] = "none, linear-gradient(to left bottom, red, blue) padding-box" should set mask-size
  assert_equals: mask-size should be canonical expected "auto, auto" but got "initial, initial"
[PASS] e.style['mask'] = "none, linear-gradient(to left bottom, red, blue) padding-box" should not set unrelated longhands
Harness: the test ran to completion.
